+++ .gitignore
@@ -12,3 +12,4 @@
 .Trashes
 ehthumbs.db
 Thumbs.db
+.phpunit.result.cache
