--- lwtv-plugin.php
@@ -1,5 +1,0 @@
-<?php
-
-
-// This is a fake file, used to make PHPUnit happy
-require_once 'functions.php';
