+++ phpcs.xml.dist
@@ -32,7 +32,7 @@
 
 	<rule ref="WordPress.Security.EscapeOutput">
 		<properties>
-			<property name="customAutoEscapedFunctions" value="LWTV_Functions,symbolicons" type="array" />
+			<property name="customAutoEscapedFunctions" value="LWTV_Features,symbolicons" type="array" />
 		</properties>
 	</rule>
