+++ theme/class-data-character.php
@@ -60,7 +60,7 @@ class LWTV_Theme_Data_Character {
 			foreach ( $lez_cliches as $the_cliche ) {
 				$termicon = get_term_meta( $the_cliche->term_id, 'lez_termsmeta_icon', true );
 				$tropicon = $termicon ? $termicon . '.svg' : 'square.svg';
-				$icon     = ( new LWTV_Functions() )->symbolicons( $tropicon, 'fa-square' );
+				$icon     = ( new LWTV_Features() )->symbolicons( $tropicon, 'fa-square' );
 				$cliches .= '<a href="' . get_term_link( $the_cliche->slug, 'lez_cliches' ) . '" data-bs-target="tooltip" data-placement="bottom" rel="tag" title="' . $the_cliche->name . '"><span role="img" aria-label="' . $the_cliche->name . '" class="character-cliche ' . $the_cliche->slug . '">' . $icon . '</span></a>&nbsp;';
 			}
 
@@ -78,7 +78,7 @@ class LWTV_Theme_Data_Character {
 	 */
 	public function dead( $character_id, $format ) {
 		$deadpage = get_term_by( 'slug', get_query_var( 'term' ), get_query_var( 'taxonomy' ) );
-		$icon     = ( new LWTV_Functions() )->symbolicons( 'grim-reaper.svg', 'fa-times-circle' );
+		$icon     = ( new LWTV_Features() )->symbolicons( 'grim-reaper.svg', 'fa-times-circle' );
 
 		// Show nothing on ARCHIVE pages for dead
 		if ( ! empty( $term ) && $format === $term->slug ) {
@@ -106,7 +106,7 @@ class LWTV_Theme_Data_Character {
 			$actor_more = ( $num_actors > 1 ) ? ' (plus ' . ( $num_actors - 1 ) . ' more)' : '';
 			$actor_post = get_post( $actor_value );
 			$actor_name = ( isset( $actor_post->post_title ) && ! is_null( $actor_post->post_title ) ) ? $actor_post->post_title : 'TBD';
-			$icon       = ( new LWTV_Functions() )->symbolicons( 'user.svg', 'fa-user' );
+			$icon       = ( new LWTV_Features() )->symbolicons( 'user.svg', 'fa-user' );
 			$output     = '<div class="card-meta-item actors">' . $icon;
 			if ( get_post_status( $actor_value ) === 'private' ) {
 				if ( is_user_logged_in() ) {
@@ -147,7 +147,7 @@ class LWTV_Theme_Data_Character {
 			$num_shows = count( $all_shows );
 			$show_more = ( $num_shows > 1 ) ? ' (plus ' . ( $num_shows - 1 ) . ' more)' : '';
 			$show_post = get_post( $shows_value['show'] );
-			$icon      = ( new LWTV_Functions() )->symbolicons( 'tv-hd.svg', 'fa-tv' );
+			$icon      = ( new LWTV_Features() )->symbolicons( 'tv-hd.svg', 'fa-tv' );
 			$output    = '<div class="card-meta-item shows">' . $icon . '<em>';
 			if ( get_post_status( $shows_value['show'] ) !== 'publish' ) {
 				$output .= '<span class="disabled-show-link">' . $show_post->post_title . '</span>';
